more defensive coding for register users
