+++ index.html
@@ -69,7 +69,7 @@
             </div>
         </main>
     </body>
-    <script src="./src/js/config.js"></script>
+    <!--script src="./src/js/config.js"></script-->
     <script src="./src/js/server-status.js"></script>
     <script src="./src/js/list-contracts.js"></script>
